# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk sistem pengelolaan museum",
        "title": "API Pengelolaan Gudang Koleksi Museum",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ghaidafasya24",
            "email": "714220031@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "inventorymuseum-de54c3e9b901.herokuapp.com",
    "basePath": "/api",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Endpoint untuk mengambil seluruh data admin yang tersimpan di sistem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Login user dan dapatkan token JWT untuk autentikasi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Data login user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Kredensial tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Registrasi akun admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterResponse"
                        }
                    },
                    "401": {
                        "description": "Username already exists",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponseRegister"
                        }
                    }
                }
            }
        },
        "/users/username/{username}": {
            "get": {
                "description": "Mengambil data detail user berdasarkan username tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User by Username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetUserByUsernameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "User tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Mengambil data user berdasarkan ID MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil ditampilkan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data user berdasarkan ID (wajib autentikasi JWT Bearer)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username (huruf kecil semua)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nomor telepon format 62xxxxxxxx",
                        "name": "phone_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password baru",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data user berdasarkan ID (wajib autentikasi JWT Bearer)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid credentials"
                }
            }
        },
        "model.ErrorResponseRegister": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Username already exists"
                }
            }
        },
...
        "model.GetAllUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Users"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Berhasil mengambil semua data users"
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.GetUserByUsernameResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Users"
                },
                "message": {
                    "type": "string",
                    "example": "User ditemukan"
                }
            }
        },
        "model.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin12345"
                },
                "username": {
                    "type": "string",
                    "example": "ghaida"
                }
            }
        },
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "expires": {
                    "type": "string",
                    "example": "2026-01-22T15:11:51.917322007Z"
                },
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "model.RegisterRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin12345"
                },
                "phone_number": {
                    "type": "string",
                    "example": "6281234567890"
                },
                "username": {
                    "type": "string",
                    "example": "ghaida"
                }
            }
        },
        "model.RegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User registered successfully"
                },
                "status": {
                    "type": "integer",
                    "example": 201
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "_id": {
                            "type": "string",
                            "example": "696ef88677f450e9430a144e"
                        },
                        "role": {
                            "type": "string",
                            "example": "admin"
                        }
                    }
                }
            }
        },
        "model.Users": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "12345678"
                },
                "phone_number": {
                    "type": "string",
                    "example": "6281234567890"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "ghaida"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003cJWT Token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}